19카메라

+++ GameEngineCore/GameEngineShaderResHelper.cpp
@@ -57,7 +57,7 @@ void GameEngineShaderResHelper::Setting()
 	}
 }
 
-void GameEngineShaderResHelper::SetConstantBufferLink(const std::string_view& _Name, const void* _Data, size_t _Size)
+void GameEngineShaderResHelper::SetConstantBufferLink(const std::string_view& _Name, const void* _Data, UINT _Size)
 {
 	std::string UpperName = GameEngineString::ToUpper(_Name);
